added colors for GEO member class, moved findDuplicates function into its own file in utils
