+++ src/App.js
@@ -1,24 +1,3 @@
-//TODOs
-/*
-refactor for the react chart.js library
-
-
-maybe need to modify the db schema
-create a table with columns for companies and company id
-join on company id in the main table
-might speed up getting the list of companies for the select menu
-and also enable more restrictive regex on the inputs for company
-  instead of having to support any company name with any
-  non-alphanumeric characters, query would just be by numeric id
-
-FEATURES TO ADD:
-  highlight text that has been filtered for OR give some better
-    indication that the results have been updated for the terms
-  replot charts for filtered text results
-  pagination for job call card results
-
-*/
-
 import React, { useEffect, useState } from 'react';
 import './App.css';
 import LineGraph from './Components/LineGraph/LineGraph';
@@ -31,6 +10,7 @@ import ClassificationPicker from './Components/ClassificationPicker/Classificati
 import CallCardList from './Components/CallCardList/CallCardList';
 import Loader from './Components/Loader/Loader';
 import handleFetch from './utils/HandleFetch';
+import findDuplicates from './utils/findDuplicates';
 import { colors } from './config';
 import { createDate } from './utils/createDate';
 
@@ -108,73 +88,9 @@ const App = () => {
     setSearchField(event.target.value);
   }
 
-  const filteredCalls = callCardData.filter(call => {
-    return call.summary.toLowerCase().includes(searchField.toLowerCase());
-  })
+  const filteredCalls = callCardData
+    .filter(call => call.summary.toLowerCase().includes(searchField.toLowerCase()));
 
-  function findDuplicates(callSheet) {
-    // console.log(callSheet);
-    //can't modify callSheet as it is a reference to callCardData
-
-    /* move through the callCardData
-    store the first union_call_id
-    check it against every other call in the array of call objects
-    on the first duplicate found, start a new duplicate object
-    insert the company, dates and the indicies of callCardData where they were found
-
-    on the next pass don't check the indexes we've already found duplicates at
-    */
-
-    let duplicates = {};
-    // array of indicies not to check since we've already found dupes there
-    let found = [];
-    let totalDuplicateMembers = 0;
-
-    for (let i = 0; i < callSheet.length; i++) {
-      if (found.includes(i)) continue;
-      let current = callSheet[i].union_call_id;
-
-      for (let j = i + 1; j < callSheet.length; j++) {
-        if (callSheet[j].union_call_id === current) {
-          if (!duplicates[current]) {
-            // start a new duplicate object
-            duplicates[current] = {
-              company: callSheet[j].company,
-              duplicate_dates: [],
-              duplicate_indicies: [],
-              duplicate_members: callSheet[j].members_needed,
-            }
-            duplicates[current]
-              .duplicate_dates
-              .push(callSheet[i].call_date_from_html, callSheet[j].call_date_from_html);
-            duplicates[current]
-              .duplicate_indicies
-              .push(i, j);
-            found.push(j);
-          } else {
-            // push our values into the exisiting duplicate object
-            duplicates[current]
-              .duplicate_dates
-              .push(callSheet[j].call_date_from_html);
-            duplicates[current]
-              .duplicate_indicies
-              .push(j);
-            duplicates[current]
-              .duplicate_members
-              += callSheet[j].members_needed;
-            totalDuplicateMembers += callSheet[j].members_needed;
-            found.push(j);
-          }
-        }
-      }
-    }
-    return ({
-      duplicates,
-      found,
-      count: found.length,
-      totalDuplicateMembers
-    });
-  }
   const staleCalls = findDuplicates(filteredCalls);
 
   return (
